Create config-updater package and extract handlers from CloudflareTunnelClient
Created new packages/config-updater package that handles configuration update endpoints, separating concerns from CloudflareTunnelClient which now focuses solely on tunnel establishment.

This is part of a refactoring effort (CYPACK-251). The CLI app has expected type errors that will be resolved in follow-up work to integrate ConfigUpdater with SharedApplicationServer.

Changes:
- Created packages/config-updater with ConfigUpdater class
- Moved all handlers from cloudflare-tunnel-client to config-updater
- ConfigUpdater registers routes with Fastify (SharedApplicationServer)
- Implements Bearer token authentication via CYRUS_API_KEY
- Handles endpoints: /api/update/cyrus-config, /api/update/cyrus-env, /api/update/repository, /api/test-mcp, /api/configure-mcp
- Simplified CloudflareTunnelClient to only manage tunnel connections
- Removed all handler-related code from CloudflareTunnelClient
- Updated CloudflareTunnelClient API: constructor now requires localPort, startTunnel() takes only cloudflareToken

Packages affected:
- packages/config-updater (new)
- packages/cloudflare-tunnel-client (refactored)

Note: CLI integration to be completed in separate work.
