+++ packages/config-updater/src/handlers/configureMcp.ts
