+++ packages/config-updater/src/index.ts
@@ -1,0 +1,2 @@
+export { ConfigUpdater } from "./ConfigUpdater.js";
+export * from "./types.js";
